Add TrackService and TrackRequirements features
New classes TrackService, TrackRequirements and Track are added to retrieve and manage playlist tracks based on user defined parameters (Instrumentalness and Speechiness). This is done by updating the Index.razor page with an edit form for submitting these parameters and a table for displaying the matching tracks. Also, unnecessary usings are removed from SpotifyService.cs.