Refactor UI and remove unnecessary import in TrackService
In TrackService, unnecessary import of SpotifyBlazor.Models has been removed to clean up the code. In Index.razor, the typos and Bootstrap layout have been improved for a better user interface. The form layout, previously structured with gy-2 gx-3 classes was changed to a simpler layout with container and row classes to enhance readability and aesthetic appeal. The typo in the playlist count message was also corrected.