WIP - Update UI design and remove redundant elements
This commit updates the design of the UI, changing font, margins, and the linear gradient for a better visual experience. Additionally, elements and code that were currently not being used or are unnecessary were removed/commented out in the files for better readability and efficiency. The NavMenu, for instance, was commented out as it was not actively contributing to the project's functionality. The Spotify page layout has been revamped with improvements in elements' arrangement and style.