Add TrackAttributteFilter component, adjust TrackService and update other files
The TrackAttributeFilter component was added to create a more modular way of setting the track attributes. This led to a significant clean up in index.razor file. The TrackService was also updated to include an event that is invoked when tracks are added. This allows the UI to react when tracks have been added. This not only aids in better code management, but also allows for an improved UI response. An additional import made in _Imports.razor to define the new component in the project namespace.